Feat Login Page works with Loading
